further increased the compatibility to opentracing
`extract` returns the original context now. So to say, the context of the parent. This is necessary because the parameter :child_of is supposed to be the context of the parent and not the context of the new span. The actually context of the new span, the child context, will be created within `prepare_span_context` from the parent context.