fix: child_duration with threaded children
We don’t realy know what we are doing here. But for know we consider the time where any child is active as `child_duration`. While the time where no child is active as `self_time`. Accounting for sequencial, nested and threaded children. If children outlife there parent, there child_duration is not accounted for.